app/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2
python-dotenv
alembic
pytest
pandas